Amend Media Owner

+++ db/migrate/20160909173742_create_media.rb
@@ -2,7 +2,7 @@ class CreateMedia < ActiveRecord::Migration[5.1]
   def change
     create_table :media do |t|
       t.string :name, :null => false
-      t.integer :owner_id, :null => false
+      t.bigint :owner_id, :null => false
       t.string :media_type, :null => false
       t.timestamp :deleted_at, :index => true
       t.string :slug, :null => false
